Install Guide: Make MEDM entry more like other DMs
In the Display Managers section of the Install Guide, make the
MEDM entry more like the rest of the entries.  Use "MEDM" instead
of "medm".  Change the links to use "epics.anl.gov" instead of
"www.aps.anl.gov/epics/".  Fix a typo in the sentence about the Windows
MSI installer.