Enhance R3-13 release notes
This includes all modules that have been updated, linked to their
corresponding release notes. When such notes are not available or
partially written, a link is provided to the change set (commit diff on
GitHub).

This makes detailed explanation of the changes more accessible to guide
users' decisions on updating.