Add release note generation script
git-diff(1) is used to track which changes were made, so that we only
check the required submodules. To take submodule changes, a match
against git-ls-tree(1) output is done to retrieve the first commit in
the submodule in the specified top-level range. An alternative (and
simpler) approach to take the same information is to use `git-diff
--submodule=log` option and parse the "Submodule" lines, but this is not
known to be stable.

The first commmit is used to get tags created since then until HEAD,
using git-tag(1) API. This simplification of using HEAD can be easily
removed, to allow arbitrary past ranges.