Generate release notes with sections
The included sections are tagged, untagged and new modules. Commits are
taken only from the diff range, so that it can also be used for
generating releases from arbitrary commit ranges.

Tagged modules are the ones that were git-tagged in the release commit
range. Such modules have well-defined release notes that document the
changes, which are linked.

On the other hand, untagged modules do not necessarily have partial
release notes, so only their changeset is linked instead.

New modules are listed in the same format used by (un)tagged modules,
but in a separate section to make clear they weren't previously
available.

A 'removed modules' section could be also created. However, an
unexpected behavior of git-diff(1) is that it does not include removed
submodules in its output. Thus, another approach should be taken for
that.